refactor: move functions to the functions.php file
