+++ add.php
@@ -2,185 +2,7 @@
 
 require_once 'init.php';
 require_once 'helpers.php';
-
-function esc (string $str)
-{
-    return htmlspecialchars($str, ENT_QUOTES);
-}
-
-function clean(string $value) {
-    $value = trim($value);
-    $value = stripslashes($value);
-    $value = strip_tags($value);
-    $value = htmlspecialchars($value);
-
-    return $value;
-}
-
-function check_text(string $value)
-{
-    $error = '';
-    if (empty($value)) {
-        $error = 'Это поле должно быть заполнено';
-    }
-    return $error;
-}
-
-function check_youtube_domain (string $value)
-{
-    $domain = parse_url($value, PHP_URL_HOST);
-    if (strpos($domain, 'youtube.com') == false) {
-        return 'Введите ссылку на видео из YOUTUBE';
-    }
-    return '';
-}
-
-function link_validate (string $link_value)
-{
-    $error = '';
-    if (empty($link_value)) {
-        $error = "Это поле должно быть заполнено";
-    } elseif (!filter_var($link_value, FILTER_VALIDATE_URL)) {
-        $error = "Значение не является ссылкой";
-    }
-    return $error;
-}
-
-function tags_validate (array $tags_value)
-{
-    $tags_error = [];
-    $tags_array = explode(' ', $tags_value);
-
-    foreach ($tags_array as $tag) {
-
-        if (!preg_match('/^[a-zA-Zа-яА-Я0-9]+$/', $tag)) {
-            $tags_error[] = $tag;
-        }
-
-       if (!empty($tags_error)) {
-           $count_error_tags = count($tags_error);
-           $tag_error = get_noun_plural_form($count_error_tags, 'Значение ', 'Значения ', 'Значения ') . implode(', ', $tags_error) . get_noun_plural_form($count_error_tags, ' не корректно', ' не корректны', ' не корректны');
-       }
-
-    }
-    return $tag_error;
-}
-
-function youtube_url_validation (string $url_value)
-{
-    $error = '';
-    if (link_validate($url_value)) {
-        $error = link_validate($url_value);
-    } elseif (check_youtube_url($url_value) !== true) {
-        $error = check_youtube_url($url_value);
-    } elseif (check_youtube_domain($url_value)) {
-        $error = check_youtube_domain($url_value);
-    }
-    return $error;
-}
-
-function check_img_type (string $file_type)
-{
-    $required_types = ['image/jpg', 'image/jpeg', 'image/png', 'image/gif'];
-    if (!in_array($file_type, $required_types)) {
-        return 'Загрузите картинку в одном из допустимых форматов: JPG PNG GIF';
-    }
-    return true;
-}
-
-function photo_validate (array $upload_photo)
-{
-    $error = '';
-    $tmp_name = $upload_photo['tmp_name'];
-    $file_info = finfo_open(FILEINFO_MIME_TYPE);
-    $file_type = finfo_file($file_info, $tmp_name);
-    if (check_img_type($file_type) !== true) {
-        $error = check_img_type($file_type);
-    }
-    return $error;
-}
-
-function photo_link_validate (string $photo_link)
-{
-    $error = '';
-    if (!empty($photo_link)) {
-        if (!filter_var($photo_link, FILTER_VALIDATE_URL)) {
-            $error = "Значение не является ссылкой";
-        } else {
-            $get_headers = get_headers($photo_link, 1);
-            if (!strpos($get_headers[0], '200')) {
-                $error = "Страница не отвечает";
-            } elseif (check_img_type($get_headers['Content-Type']) !== true) {
-                $error = "Ссылка на недопустимый тип файла";
-            }
-        }
-    }
-    return $error;
-}
-
-function upload_photo (array $upload_photo)
-{
-    $tmp_name = $upload_photo['tmp_name'];
-    $path = $upload_photo['name'];
-    $filename = 'uploads/' . uniqid() . '.' . pathinfo($path, PATHINFO_EXTENSION);
-    move_uploaded_file($tmp_name, $filename);
-    return $filename;
-}
-
-function is_new_tag (string $tag, array $existing_tags)
-{
-    return !(in_array($tag, $existing_tags));
-}
-
-function rules($post_type, $tags)
-{
-    $rules = [];
-    $rules['title'] = function ($value) {
-        return check_text($value);
-    };
-    if (!empty($tags)) {
-        $rules['tags'] = function ($value) {
-            return tags_validate($value);
-        };
-    }
-    switch ($post_type) {
-        case 'text':
-            $rules['content'] = function ($value) {
-                return check_text($value);
-            };
-            break;
-        case 'link':
-            $rules['link'] = function ($value) {
-                return link_validate($value);
-            };
-            break;
-        case 'video':
-            $rules['video'] = function ($value) {
-                return youtube_url_validation($value);
-            };
-            break;
-        case 'quote':
-            $rules['content'] = function ($value) {
-                return check_text($value);
-            };
-            $rules['author_quote'] = function ($value) {
-                return check_text($value);
-            };
-            break;
-        case 'photo' :
-            if (!empty($_FILES['upload_photo']['name'])) {
-                $rules['photo'] = function ($value) {
-                    return photo_validate($value);
-                };
-            } else {
-                $rules['img'] = function ($value) {
-                    return photo_link_validate($value);
-                };
-            }
-            break;
-    }
-    return $rules;
-}
+require_once 'functions.php';
 
 if (!$link) {
     print 'error' . mysqli_connect_error();
